+++ src/i18n/messages/es-MX.ts
@@ -65,11 +65,13 @@ export const Spanish: Messages = {
   Settings7: "Idioma",
   Settings8: "Próximamente Globle Edición de Ciudades",
   Settings9: "Practicar",
+  Settings10: "Arco iris activado",
+  Settings11: "Arco iris desactivado",
   Answer: "Respuesta",
   Closest: "Los países más cercanos",
   PracticeMode: "You are in practice mode.", //TODO: Translate
   PracticeExit: "Exit practice mode", //TODO: Translate
-  PracticeNew: "New practice game" //TODO: Translate
+  PracticeNew: "New practice game", //TODO: Translate
 };
 
 // export default {