+++ src/i18n/messages/pt-BR.ts
@@ -68,9 +68,11 @@ export const Portuguese: Messages = {
   Settings7: "Idioma",
   Settings8: "Globle: Edição Cidades em breve!",
   Settings9: "Praticar",
+  Settings10: "Arco-íris ativado",
+  Settings11: "Arco-íris desativado",
   Answer: "Answer", //TODO: Translate
   Closest: "Closest", //TODO: Translate
   PracticeMode: "You are in practice mode.", //TODO: Translate
   PracticeExit: "Exit practice mode", //TODO: Translate
-  PracticeNew: "New practice game" //TODO: Translate
+  PracticeNew: "New practice game", //TODO: Translate
 };